refactor(app): 🎨 refactor getThouths
