--- apps/web/src/messages/index.ts
@@ -1,2 +1,0 @@
-export const CONFIRM_MESSAGE = 'confirm delete account'
-export const USER_NOT_FOUND_ERROR = "The user couldn't be found, try again."
