+++ apps/web/src/types/index.d.ts
@@ -1,0 +1,5 @@
+export interface ActionResponse<T> {
+  status: number
+  data: T
+  message?: string
+}
