+++ apps/web/tsconfig.json
@@ -15,11 +15,11 @@
         "src/components/ui/*",
         "src/components/*"
       ],
+      "@lib/*": [
+        "src/lib/*"
+      ],
       "@/*": [
         "src/*",
-      ],
-      "@lib/*": [
-        "src/*"
       ],
       "contentlayer/generated": [
         "./.contentlayer/generated"
