+++ packages/utils/src/messages/index.ts
@@ -1,0 +1,8 @@
+// ERRORS
+export const ERROR_LOGIN_REQUIRED = 'You must be logged in.'
+
+export const CONFIRM_MESSAGE = 'confirm delete account'
+export const USER_NOT_FOUND_ERROR = "The user couldn't be found, try again."
+
+// THOUGHTS
+export const NOT_FOUND_THOUGHTS = 'Not found thoughts'
